% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico_temp_mensual.R
\name{grafico_temperatura_mensual}
\alias{grafico_temperatura_mensual}
\title{gráfico de la temperatura mensual}
\usage{
grafico_temperatura_mensual(
  estaciones,
  colores = c("darkgreen", "lightblue", "darkred", "pink", "purple"),
  titulo = "Temperatura Mensual"
)
}
\arguments{
\item{estaciones}{dataframe con columnas "fecha", "temperatura_abrigo_150cm" y "id"}

\item{colores}{vector de colores para las series (opcional)}

\item{titulo}{título del gráfico. Por defecto "Temperatura"}
}
\value{
un objeto ggplot
}
\description{
grafica la temperatura mensual por estación
}
\examples{
estaciones <- tibble::tibble(
  id  = c("A","A","B","B"),
  fecha = as.Date(c("2020-01-01","2020-02-01","2020-01-01","2020-02-01")),
  temperatura_abrigo_150cm = c(10, 12, 15, 11)
)
grafico_temperatura_mensual(estaciones)

}
