% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leer_estaciones.R
\name{leer_estacion}
\alias{leer_estacion}
\title{lee datos de estaciones}
\usage{
leer_estacion(id_estacion, ruta)
}
\arguments{
\item{id_estacion}{código de la estación}

\item{ruta}{ruta local donde guardar/leer el CSV}
}
\value{
un dataframe con los datos de la estación
}
\description{
descarga y lee los datos de estaciones
}
\examples{
leer_estacion("NH0437", tempfile(fileext = ".csv"))

}
