% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabla_resumen_temp.R
\name{tabla_resumen_temperatura}
\alias{tabla_resumen_temperatura}
\title{tabla resumen de temperatura}
\usage{
tabla_resumen_temperatura(datos)
}
\arguments{
\item{datos}{dataframe con las columnas "id" y "temperatura_abrigo_150cm"}
}
\value{
un tibble con promedio y desvío estándar por estación
}
\description{
devuelve una tabla de resumen de temperatura_abrigo_150cm para una o más estaciones.
}
\examples{
datos <- tibble::tibble(
  id  = c("A","A","B","B"),
  fecha = as.Date(c("2020-01-01","2020-02-01","2020-01-01","2020-02-01")),
  temperatura_abrigo_150cm = c(10, 12, 15, 11)
)
tabla_resumen_temperatura(datos)

}
